// src/types.ts
export type PlayResultRank =
  | "D"
  | "C"
  | "B"
  | "BB"
  | "BBB"
  | "A"
  | "AA"
  | "AAA"
  | "S"
  | "S+"
  | "SS"
  | "SS+"
  | "SSS"
  | "SSS+";

export type RatingType = "best" | "recent" | "selection";

export type StdChartDifficulty =
  | "basic"
  | "advanced"
  | "expert"
  | "master"
  | "ultima";